requirements.txt
djangorestframework
django
firebase-admin
djangorestframework-simplejwt